+++ package.json
@@ -19,6 +19,7 @@
 		"@prisma/extension-accelerate": "^1.2.1",
 		"@radix-ui/react-alert-dialog": "^1.1.4",
 		"@radix-ui/react-checkbox": "^1.1.2",
+		"@radix-ui/react-context-menu": "^2.2.4",
 		"@radix-ui/react-dialog": "^1.1.2",
 		"@radix-ui/react-dropdown-menu": "^2.1.3",
 		"@radix-ui/react-label": "^2.1.0",
